services/ml/requirements.txt
# ML Training Pipeline Dependencies

# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Deep learning
tensorflow==2.13.0
tensorflowjs==4.10.0

# Data processing
scipy==1.11.1
imbalanced-learn==0.11.0

# Model persistence
joblib==1.3.2

# PostgreSQL connectivity
psycopg2-binary==2.9.7
sqlalchemy==2.0.19

# Experiment tracking
mlflow==2.6.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
